// frontend/src/models/types.d.ts
export interface ProfileResponse {
  address_id?: string;
  address_robust?: string;
  avatar: string;
  contact_email: string;
  description: string;
  github: string;
  location: string;
  logo: string;
  name: string;
  reputation_score?: number;
  review_score?: number;
  reviews?: number;
  slack: string;
  twitter: string;
  type?: string;
  uid?: string;
}